refactor(app): update types, utils, constants
